Add js modules
In this commit the js modules were added
