+++ js/burger-menu.js
@@ -1,0 +1,11 @@
+//burger-con
+
+export const initializeHamburgerMenu = () => {
+const hamburger = document.querySelector('.hamburger');
+const navMenu = document.querySelector('.nav-menu');
+
+hamburger.addEventListener("click", () => {
+    hamburger.classList.toggle("active");
+    navMenu.classList.toggle("active");
+});
+};
