+++ js/video-player.js
@@ -1,0 +1,4 @@
+//video
+export const initializeVideoPlayer = () => {
+    const player = new Plyr('video');
+};
