+++ projects.php
@@ -29,7 +29,7 @@
     <script defer src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.5.0/gsap.min.js"></script>
     <script defer src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.5.0/ScrollTrigger.js"></script>
     <script defer src="https://cdn.plyr.io/3.7.8/plyr.js"></script>
-    <script defer src="js/main.js"></script>
+    <script type="module" defer src="js/main.js"></script>
 </head>
 
 <body>
@@ -89,11 +89,11 @@
 
         <section class="grid-con projects-section-projectcards">
             <h2 class="hidden">Project Cards</h2>
-            <div class="filter-buttons col-span-full">
+            <!-- <div class="filter-buttons col-span-full">
                 <p class="filter-button">All</p>
                 <p class="filter-button">Web Development</p>
                 <p class="filter-button">3D Motion Design</p>
-            </div>
+            </div> -->
 
             <?php
