+++ sass/pages/_home.scss
@@ -213,24 +213,26 @@ section {
     width: 400px;
     background: white;
     border-radius: 6px;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
+    margin: 100px auto;
     text-align: center;
     padding: 0 30px 30px;
     color: #333;
     border-radius: 15px;
+
+    h2 {
+        font-size: 38px;
+        font-weight: bold;
+        margin: 30px 0 10px;
+        color: black;
+        }
+    
+        p {
+            color: black;
+        }
 }
 
 .popup-message img {
     width: 150px;
-}
-
-.popup-message h2 {
-    font-size: 38px;
-    font-weight: bold;
-    margin: 30px 0 10px;
 }
 
 .button-ok{
